Add prompt for removal of data volume

+++ clear_up.sh
@@ -23,8 +23,13 @@ docker image rm $JENKINS_IMAGE_NAME
 echo "<---Removing docker network--->"
 docker network rm $NETWORK_NAME
 
-echo "<---Removing data volumes--->"
-docker volume rm $DATA_VOLUME_NAME
+read -p "Remove data volume? (y/n): " remove_data_volume
+if [[ "$remove_data_volume" == "y" ]]; then
+  echo "<---Removing data volume--->"
+  docker volume rm $DATA_VOLUME_NAME
+fi
+
+echo "<---Removing certs volume--->"
 docker volume rm $CERTS_VOLUME_NAME
 
 echo "<---Script Complete--->"
